+++ modules/AWSPowerShell/Cmdlets/MigrationHubRefactorSpaces/Basic/New-MHRSApplication-Cmdlet.cs
@@ -31,8 +31,8 @@ namespace Amazon.PowerShell.Cmdlets.MHRS
     /// Creates an Amazon Web Services Migration Hub Refactor Spaces application. The account
     /// that owns the environment also owns the applications created inside the environment,
     /// regardless of the account that creates the application. Refactor Spaces provisions
-    /// the Amazon API Gateway and Network Load Balancer for the application proxy inside
-    /// your account.
+    /// an Amazon API Gateway, API Gateway VPC link, and Network Load Balancer for the application
+    /// proxy inside your account.
     /// </summary>
     [Cmdlet("New", "MHRSApplication", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
     [OutputType("Amazon.MigrationHubRefactorSpaces.Model.CreateApplicationResponse")]